+++ webgear-openapi/src/WebGear/OpenApi/Trait/Cookie.hs
@@ -12,12 +12,13 @@ import WebGear.Core.Request (Request)
 import WebGear.Core.Response (Response)
 import WebGear.Core.Trait (Get (..), Set (..), Trait, TraitAbsence)
 import qualified WebGear.Core.Trait.Cookie as WG
-import WebGear.OpenApi.Handler (DocNode (..), OpenApiHandler (..), nullNode, singletonNode)
+import WebGear.OpenApi.Handler (OpenApiHandler (..))
+import WebGear.OpenApi.Trait.Auth (addSecurityScheme)
 
 instance (KnownSymbol name, TraitAbsence (WG.Cookie e name val) Request) => Get (OpenApiHandler m) (WG.Cookie e name val) Request where
   {-# INLINE getTrait #-}
   getTrait WG.Cookie =
-    OpenApiHandler $ singletonNode $ DocSecurityScheme cookieName securityScheme
+    OpenApiHandler $ addSecurityScheme cookieName securityScheme
     where
       cookieName = fromString @Text $ symbolVal $ Proxy @name
 
@@ -37,4 +38,4 @@ instance (KnownSymbol name, TraitAbsence (WG.Cookie e name val) Request) => Get 
 
 instance (Trait (WG.SetCookie e name) Response) => Set (OpenApiHandler m) (WG.SetCookie e name) Response where
   {-# INLINE setTrait #-}
-  setTrait WG.SetCookie _ = OpenApiHandler nullNode
+  setTrait WG.SetCookie _ = OpenApiHandler pure