+++ webgear-openapi/src/WebGear/OpenApi/Trait/QueryParam.hs
@@ -3,21 +3,28 @@
 -- | OpenApi implementation of 'QueryParam' trait.
 module WebGear.OpenApi.Trait.QueryParam where
 
+import Control.Lens ((&), (.~), (<>~))
+import Control.Monad.State.Strict (MonadState)
 import Data.OpenApi (
+  OpenApi,
   Param (..),
   ParamLocation (ParamQuery),
   Referenced (Inline),
   ToSchema,
+  allOperations,
+  description,
+  parameters,
   toSchema,
  )
 import Data.Proxy (Proxy (Proxy))
 import Data.String (fromString)
 import GHC.TypeLits (KnownSymbol, symbolVal)
-import WebGear.Core.Modifiers
+import WebGear.Core.Handler (Description (..))
+import WebGear.Core.Modifiers (Existence (..))
 import WebGear.Core.Request (Request)
 import WebGear.Core.Trait (Get (..), TraitAbsence)
 import WebGear.Core.Trait.QueryParam (QueryParam (..))
-import WebGear.OpenApi.Handler (DocNode (DocQueryParam), OpenApiHandler (..), singletonNode)
+import WebGear.OpenApi.Handler (Documentation (..), OpenApiHandler (..), consumeDescription)
 
 instance (KnownSymbol name, ToSchema val, TraitAbsence (QueryParam Required ps name val) Request) => Get (OpenApiHandler m) (QueryParam Required ps name val) Request where
   {-# INLINE getTrait #-}
@@ -29,7 +36,7 @@ instance (KnownSymbol name, ToSchema val, TraitAbsence (QueryParam Required ps n
             , _paramRequired = Just True
             , _paramSchema = Just $ Inline $ toSchema $ Proxy @val
             }
-     in OpenApiHandler $ singletonNode (DocQueryParam param)
+     in OpenApiHandler $ addParam param
 
 instance (KnownSymbol name, ToSchema val, TraitAbsence (QueryParam Optional ps name val) Request) => Get (OpenApiHandler m) (QueryParam Optional ps name val) Request where
   {-# INLINE getTrait #-}
@@ -41,4 +48,10 @@ instance (KnownSymbol name, ToSchema val, TraitAbsence (QueryParam Optional ps n
             , _paramRequired = Just False
             , _paramSchema = Just $ Inline $ toSchema $ Proxy @val
             }
-     in OpenApiHandler $ singletonNode (DocQueryParam param)
+     in OpenApiHandler $ addParam param
+
+addParam :: (MonadState Documentation m) => Param -> OpenApi -> m OpenApi
+addParam param doc = do
+  desc <- consumeDescription
+  let param' = param & description .~ fmap getDescription desc
+  pure $ doc & allOperations . parameters <>~ [Inline param']